refactor: complete appointmentNote refactoring and add pre-push check scripts
## Refactoring (690 → 189 lines, 73% reduction)
- Created modular architecture with clear separation of concerns:
  - API service layer (appointmentNote.service.ts)
  - UI components (Header, EditDialog, ErrorStates)
  - Custom hooks for data fetching (useAppointmentData)
  - Custom hooks for mutations (useAppointmentNoteMutations)
  - Custom hooks for handlers (useNoteHandlers)
  - Custom hooks for templates (useAppointmentTemplates)
- Fixed all TypeScript errors and linting issues
- Removed debug console.log statement
- Maintained all original functionality

## Pre-push Check Scripts
- Created pre-push-check.sh: Comprehensive script that runs all CI/CD checks locally
- Created local-check-test.sh: Interactive script for quick local verification
- Added npm scripts for easy access: npm run check:local and npm run check:pre-push
- Scripts automatically use local .env for integration tests
- Prevents failed pushes by catching issues early

## Benefits
- Improved code maintainability and testability
- Faster development with local check scripts
- Reduced CI/CD failures
- Better developer experience
